requirements.txt
uvicorn
langgraph
langchain
langchain_openai
langchain_groq
langchain_huggingface
langchain-community
langchain_text_splitters
beautifulsoup4
playwright
tiktoken
pydantic
supabase
fastapi